break out Gallery filter components into separate file
